controllerの修正

+++ app/controllers/items_controller.rb
@@ -40,8 +40,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.user == current_user
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to items_path
+    end
   end
 
   private
